Update test imports and type for MAX_ACTIVE_VALIDATORS
Changed import path for test_utils in view_unit_tests.rs to use crate::unit. Updated MAX_ACTIVE_VALIDATORS type from u64 to usize in test_utils.rs for consistency. Minor formatting and logic fixes in process_claims_tests.rs for delegation calculations.

+++ contracts/vault/tests/test_utils.rs
@@ -18,7 +18,7 @@ const FT_DECIMALS: u8 = 6;
 pub const VAULT_CALL_GAS: Gas = Gas::from_tgas(300);
 pub const YOCTO_NEAR: u128 = 10u128.pow(24);
 pub const MAX_COUNTER_OFFERS: u64 = 7;
-pub const MAX_ACTIVE_VALIDATORS: u64 = 2;
+pub const MAX_ACTIVE_VALIDATORS: usize = 2;
 
 pub struct InstantiateTestVaultResult {
     pub execution_result: ExecutionFinalResult,
